Resolved annoying warning conversion types using static_cast to desired type

+++ Project/Source/GameplaySystems.cpp
@@ -177,11 +177,11 @@ void SceneManager::ExitGame() {
 }
 
 float Screen::GetScreenWitdh() {
-	return App->window->GetWidth();
+	return static_cast<float>(App->window->GetWidth());
 }
 
 float Screen::GetScreenHeight() {
-	return App->window->GetHeight();
+	return static_cast<float>(App->window->GetHeight());
 }
 
 GameObject* Physics::Raycast(const float3& start, const float3& end) {
